requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
google-generativeai>=0.3.0
pinecone
python-dotenv>=1.0.0
pydantic>=2.3.0
numpy>=1.24.3
transformers>=4.32.1
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
nltk>=3.8.1
spacy>=3.6.1
pytest>=7.4.0
httpx>=0.24.1
python-multipart>=0.0.6
pymongo
langchain>=0.2.1
langchain-community>=0.2.1
langchain-core>=0.1.0
boto3
python-multipart
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.1.2
langchain-text-splitters
filetype
langchain_pinecone
langchain-huggingface>=0.1.2
pypdf
hf_xet
huggingface_hub[hf_xet]
unstructured>=0.5.0  # Added for UnstructuredPDFLoader
PyMuPDF>=1.21.0
pdfminer.six>=20221105
pillow-heif>=0.12.0
pillow-heif>=0.12.0
pytesseract>=0.3.10
onnxruntime
unstructured[ocr,pdf]>=0.11.0
langchain_openai
openai
cohere
-e .
